bestbuy-clearance-monitor/requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
pytest==8.3.2
python-dotenv==1.0.0
